Add files via upload
This Jupyter Notebook shows the process of two text processing packages (TextBlob and Vader) giving their predictions of whether the YT comments in the dataset are positive, negative, or neutral, compared to the original sentiment labels the comments were given.  I manipulated the original tabular data to have the results of both of the packages and saved it as a csv for my teammate Julius (KBAB) to use.